Lock integer dtype to int32 end-to-end
Tighten Int1D/IntND/DiscreteState/DiscreteAction/ScalarInt to Int32 in
typing.py, and cast searchsorted/argmax/unravel_index/where outputs to
int32 at every site where their width depended on jax_enable_x64. This
prevents the JIT cache from silently splitting into per-period
int32/int64 variants and breaks the AOT-compiled simulate program that
ships a single signature. Adds a regression test asserting discrete
grids, build_initial_states discrete entries, and MISSING_CAT_CODE
match int32.

+++ src/lcm/grids/coordinates.py
@@ -188,7 +188,7 @@ def get_irreg_coordinate(
     n_points = len(points)
 
     # Find the index of the first point greater than value
-    idx_upper = jnp.searchsorted(points, value, side="right")
+    idx_upper = jnp.searchsorted(points, value, side="right").astype(jnp.int32)
 
     # Clamp to valid range for interpolation
     idx_upper = jnp.clip(idx_upper, 1, n_points - 1)
